did I mention how much I love tests?
