+++ openpnm/algorithms/__init__.py
@@ -1,13 +1,3 @@
-r"""
-Collection of pre-defined algorithms
-====================================
-
-The ``algorithms`` module contains classes for conducting transport
-simulations on pore networks.
-
-"""
-from ._mixins import *
-
 from ._generic_algorithm import *
 from ._generic_transport import *
